Switch to regular backend and use new user id
