+++ app/src/main/java/com/cornellappdev/android/pollo/Polls/PollsChoiceRecyclerAdapter.kt
@@ -15,7 +15,7 @@ import kotlin.math.roundToInt
 
 
 class PollsChoiceRecyclerAdapter(private val poll: Poll,
-                                 private val googleId: String,
+                                 private val userId: String,
                                  private val role: User.Role) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
     private var positionSelected = -1
@@ -33,12 +33,12 @@ class PollsChoiceRecyclerAdapter(private val poll: Poll,
 
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         val choiceHolder = holder as ChoiceHolder
-        choiceHolder.bindPoll(poll, googleId, role)
+        choiceHolder.bindPoll(poll, userId, role)
         if (poll.state == PollState.live && role == User.Role.MEMBER) {
             choiceHolder.view.setOnClickListener { _ ->
                 positionSelected = position
                 choiceHolder.view.answerButton.isChecked = true
-                poll.userAnswers[googleId] = arrayListOf(position)
+                poll.userAnswers[userId] = arrayListOf(position)
                 notifyDataSetChanged()
                 sendAnswer(position)
             }
