+++ app/src/main/java/com/cornellappdev/android/pollo/models/User.java
@@ -18,6 +18,8 @@ public class User {
         this.netID = netID;
     }
 
+    public String getId() { return id; }
+
     public String getNetID() { return netID; }
 
     public enum Role {
