+++ app/src/main/java/com/cornellappdev/android/pollo/networking/Socket.kt
@@ -102,8 +102,7 @@ object Socket {
     fun connect(id: String, accessToken: String) {
         val urlBuilder = HttpUrl.Builder()
                 .scheme("https")
-                .host(BuildConfig.TEMP_BACKEND_URI)
-//                .host(BuildConfig.BACKEND_URI)
+                .host(BuildConfig.BACKEND_URI)
                 .build()
 
         Log.d("Socket URL", urlBuilder.url().toString())
